+++ crates/ruma-common/CHANGELOG.md
@@ -18,7 +18,7 @@ Improvements:
   `SystemTime` that works under WASM in the
   `MilliSecondsSinceUnixEpoch::to_system_time()` method.
 - Stabilize support for `.m.rule.suppress_edits` push rule (MSC3958 / Matrix 1.9)
-- Add `MatrixVersion::V1_9`
+- Add `MatrixVersion::V1_9` and `V1_10`
 - Point links to the Matrix 1.9 specification
 - Implement `as_str()` and `AsRef<str>` for `push::PredefinedRuleId`
 - Implement `kind()` for `push::Predefined{*}RuleId`